refactor integrationtest structure
