--- integrationtest/Makefile
@@ -1,7 +1,0 @@
-BYBIT_TEST_UPDATED=false
-
-test:
-	BYBIT_TEST_UPDATED=${BYBIT_TEST_UPDATED} BYBIT_TEST_KEY=${BYBIT_TEST_KEY} BYBIT_TEST_SECRET=${BYBIT_TEST_SECRET} go test github.com/hirokisan/bybit/integrationtest -v -tags=integrationtest
-
-test-spec:
-	BYBIT_TEST_UPDATED=${BYBIT_TEST_UPDATED} BYBIT_TEST_KEY=${BYBIT_TEST_KEY} BYBIT_TEST_SECRET=${BYBIT_TEST_SECRET} go test github.com/hirokisan/bybit/integrationtest -v -tags=integrationtest -run ${BYBIT_TEST_METHOD}
