+++ integrationtest/future-inverse-future/testdata/v2-public-account-ratio.json
