+++ integrationtest/future-inverse-future/testdata/v2-public-big-deal.json
