+++ integrationtest/future-inverse-future/testdata/v2-public-index-price-kline.json
