+++ integrationtest/future-inverse-future/testdata/v2-public-kline-list.json
