+++ integrationtest/future-inverse-future/testdata/v2-public-open-interest.json
