+++ integrationtest/future-inverse-future/testdata/v2-public-order-book-l2.json
