+++ integrationtest/future-inverse-future/testdata/v2-public-tickers.json
