+++ integrationtest/future-inverse-perpetual/Makefile
@@ -1,0 +1,7 @@
+BYBIT_TEST_UPDATED=false
+
+test:
+	BYBIT_TEST_UPDATED=${BYBIT_TEST_UPDATED} BYBIT_TEST_KEY=${BYBIT_TEST_KEY} BYBIT_TEST_SECRET=${BYBIT_TEST_SECRET} go test github.com/hirokisan/bybit/integrationtest/future-inverse-perpetual -v -tags=integrationtestfutureinverseperpetual
+
+test-spec:
+	BYBIT_TEST_UPDATED=${BYBIT_TEST_UPDATED} BYBIT_TEST_KEY=${BYBIT_TEST_KEY} BYBIT_TEST_SECRET=${BYBIT_TEST_SECRET} go test github.com/hirokisan/bybit/integrationtest/future-inverse-perpetual -v -tags=integrationtestfutureinverseperpetual -run ${BYBIT_TEST_METHOD}
