+++ integrationtest/future-inverse-perpetual/future_inverse_perpetual_test.go
@@ -1,0 +1,199 @@
+//go:build integrationtestfutureinverseperpetual
+
+package integrationtestfutureinverseperpetual
+
+import (
+	"testing"
+	"time"
+
+	"github.com/hirokisan/bybit"
+	"github.com/hirokisan/bybit/integrationtest/testhelper"
+	"github.com/stretchr/testify/require"
+)
+
+func TestBalance(t *testing.T) {
+	t.Run("ok", func(t *testing.T) {
+		client := bybit.NewTestClient().WithAuthFromEnv()
+		res, err := client.Future().InversePerpetual.Balance(bybit.CoinUSDT)
+		{
+			require.NoError(t, err)
+		}
+		{
+			goldenFilename := "./testdata/v2-private-wallet-balance.json"
+			testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+			testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		}
+	})
+
+	t.Run("auth error", func(t *testing.T) {
+		client := bybit.NewTestClient()
+		_, err := client.Future().InversePerpetual.Balance(bybit.CoinBTC)
+		require.Error(t, err)
+	})
+}
+
+func TestOrderBook(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.OrderBook(bybit.SymbolInverseBTCUSD)
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-order-book-l2.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestListKline(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.ListKline(bybit.ListKlineParam{
+		Symbol:   bybit.SymbolInverseBTCUSD,
+		Interval: bybit.Interval120,
+		From:     int(time.Now().AddDate(0, 0, -1).Unix()),
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-kline-list.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestTickers(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.Tickers(bybit.SymbolInverseBTCUSD)
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-tickers.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestTradingRecords(t *testing.T) {
+	client := bybit.NewTestClient()
+	limit := 10
+	res, err := client.Future().InversePerpetual.TradingRecords(bybit.TradingRecordsParam{
+		Symbol: bybit.SymbolInverseBTCUSD,
+		Limit:  &limit,
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-trading-records.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestSymbols(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.Symbols()
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-symbols.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestIndexPriceKline(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.IndexPriceKline(bybit.IndexPriceKlineParam{
+		Symbol:   bybit.SymbolInverseBTCUSD,
+		Interval: bybit.IntervalD,
+		From:     int(time.Now().AddDate(0, 0, -1).Unix()),
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-index-price-kline.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestOpenInterest(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.OpenInterest(bybit.OpenInterestParam{
+		Symbol: bybit.SymbolInverseBTCUSD,
+		Period: bybit.Period1h,
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-open-interest.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestBigDeal(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.BigDeal(bybit.BigDealParam{
+		Symbol: bybit.SymbolInverseBTCUSD,
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-big-deal.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestAccountRatio(t *testing.T) {
+	client := bybit.NewTestClient()
+	limit := 10
+	res, err := client.Future().InversePerpetual.AccountRatio(bybit.AccountRatioParam{
+		Symbol: bybit.SymbolInverseBTCUSD,
+		Period: bybit.Period1h,
+		Limit:  &limit,
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-account-ratio.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
+
+func TestPremiumIndexKline(t *testing.T) {
+	client := bybit.NewTestClient()
+	res, err := client.Future().InversePerpetual.PremiumIndexKline(bybit.PremiumIndexKlineParam{
+		Symbol:   bybit.SymbolInverseBTCUSD,
+		Interval: bybit.Interval120,
+		From:     int(time.Now().AddDate(0, 0, -1).Unix()),
+	})
+	{
+		require.NoError(t, err)
+		require.Equal(t, "OK", res.RetMsg)
+	}
+	{
+		goldenFilename := "./testdata/v2-public-premium-index-kline.json"
+		testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+		testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
+	}
+}
