+++ integrationtest/future-inverse-perpetual/testdata/v2-public-big-deal.json
@@ -1,0 +1,224 @@
+[
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637936558,
+    "value": 630399
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637936557,
+    "value": 531767
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637935987,
+    "value": 715445
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637928854,
+    "value": 1217800
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637928491,
+    "value": 8504430
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637927810,
+    "value": 503460
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637927810,
+    "value": 667013
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637927810,
+    "value": 599411
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637925205,
+    "value": 1000026
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637925131,
+    "value": 2579377
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637925129,
+    "value": 1300000
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637924877,
+    "value": 527817
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637924877,
+    "value": 3713815
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637924449,
+    "value": 7141905
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637916588,
+    "value": 528070
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637915876,
+    "value": 12000000
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637915873,
+    "value": 8000000
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637915873,
+    "value": 5500000
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637914900,
+    "value": 8021448
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637914898,
+    "value": 599569
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637914045,
+    "value": 690737
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637901496,
+    "value": 597981
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637895406,
+    "value": 3974001
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Sell",
+    "timestamp": 1637893718,
+    "value": 3512761
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637885187,
+    "value": 922151
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637881769,
+    "value": 500000
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637871466,
+    "value": 542352
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637862649,
+    "value": 599569
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637860556,
+    "value": 669537
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637857539,
+    "value": 538827
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637856211,
+    "value": 3000000
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637855895,
+    "value": 534260
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637855845,
+    "value": 925729
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637855834,
+    "value": 869401
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637855831,
+    "value": 866201
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637855828,
+    "value": 863788
+  },
+  {
+    "symbol": "BTCUSD",
+    "side": "Buy",
+    "timestamp": 1637852178,
+    "value": 592593
+  }
+]
