+++ integrationtest/future-inverse-perpetual/testdata/v2-public-index-price-kline.json
@@ -1,0 +1,11 @@
+[
+  {
+    "symbol": "BTCUSD",
+    "period": "D",
+    "open_time": 1637884800,
+    "open": "58961.63",
+    "high": "59193.9",
+    "low": "53546.82",
+    "close": "54712.74"
+  }
+]
