+++ integrationtest/future-inverse-perpetual/testdata/v2-public-open-interest.json
@@ -1,0 +1,252 @@
+[
+  {
+    "open_interest": 362574938,
+    "timestamp": 1637935200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 363814340,
+    "timestamp": 1637931600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 377328800,
+    "timestamp": 1637928000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 398137724,
+    "timestamp": 1637924400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 397790418,
+    "timestamp": 1637920800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 388058487,
+    "timestamp": 1637917200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 424043887,
+    "timestamp": 1637913600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 415155871,
+    "timestamp": 1637910000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 416313738,
+    "timestamp": 1637906400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 407580363,
+    "timestamp": 1637902800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 409026751,
+    "timestamp": 1637899200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 408806648,
+    "timestamp": 1637895600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 417375907,
+    "timestamp": 1637892000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 417317849,
+    "timestamp": 1637888400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 411801032,
+    "timestamp": 1637884800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 410387649,
+    "timestamp": 1637881200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 413459658,
+    "timestamp": 1637877600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 419292371,
+    "timestamp": 1637874000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 410014830,
+    "timestamp": 1637870400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 414276301,
+    "timestamp": 1637866800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 413839354,
+    "timestamp": 1637863200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 411058766,
+    "timestamp": 1637859600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 416016970,
+    "timestamp": 1637856000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 409501556,
+    "timestamp": 1637852400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 407661044,
+    "timestamp": 1637848800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 405239782,
+    "timestamp": 1637845200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 400749455,
+    "timestamp": 1637841600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 399964592,
+    "timestamp": 1637838000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 391534357,
+    "timestamp": 1637834400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 387709931,
+    "timestamp": 1637830800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 384302798,
+    "timestamp": 1637827200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 384361083,
+    "timestamp": 1637823600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 383654925,
+    "timestamp": 1637820000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 383607349,
+    "timestamp": 1637816400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 385468382,
+    "timestamp": 1637812800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 382070055,
+    "timestamp": 1637809200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 387366246,
+    "timestamp": 1637805600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 384791668,
+    "timestamp": 1637802000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 384395931,
+    "timestamp": 1637798400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 385230358,
+    "timestamp": 1637794800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 383774823,
+    "timestamp": 1637791200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 382825623,
+    "timestamp": 1637787600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 397042051,
+    "timestamp": 1637784000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 393147782,
+    "timestamp": 1637780400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 387627208,
+    "timestamp": 1637776800,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 386070934,
+    "timestamp": 1637773200,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 384566912,
+    "timestamp": 1637769600,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 377496368,
+    "timestamp": 1637766000,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 374403795,
+    "timestamp": 1637762400,
+    "symbol": "BTCUSD"
+  },
+  {
+    "open_interest": 375703730,
+    "timestamp": 1637758800,
+    "symbol": "BTCUSD"
+  }
+]
