+++ integrationtest/future-inverse-perpetual/testdata/v2-public-premium-index-kline.json
