+++ integrationtest/future-inverse-perpetual/testdata/v2-public-trading-records.json
@@ -1,0 +1,82 @@
+[
+  {
+    "id": 47946617,
+    "symbol": "BTCUSD",
+    "price": 54703,
+    "qty": 3300,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:40.52Z"
+  },
+  {
+    "id": 47946616,
+    "symbol": "BTCUSD",
+    "price": 54729.5,
+    "qty": 18,
+    "side": "Sell",
+    "time": "2021-11-26T14:41:31.111Z"
+  },
+  {
+    "id": 47946615,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.884Z"
+  },
+  {
+    "id": 47946614,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.781Z"
+  },
+  {
+    "id": 47946613,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.679Z"
+  },
+  {
+    "id": 47946612,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.574Z"
+  },
+  {
+    "id": 47946611,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.457Z"
+  },
+  {
+    "id": 47946610,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.344Z"
+  },
+  {
+    "id": 47946609,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.243Z"
+  },
+  {
+    "id": 47946608,
+    "symbol": "BTCUSD",
+    "price": 54752,
+    "qty": 100,
+    "side": "Buy",
+    "time": "2021-11-26T14:41:25.14Z"
+  }
+]
