--- integrationtest/future_inverse_perpetual_test.go
@@ -1,32 +1,0 @@
-//go:build integrationtest
-
-package integrationtest
-
-import (
-	"testing"
-
-	"github.com/hirokisan/bybit"
-	"github.com/hirokisan/bybit/integrationtest/testhelper"
-	"github.com/stretchr/testify/require"
-)
-
-func TestBalance(t *testing.T) {
-	t.Run("ok", func(t *testing.T) {
-		client := bybit.NewTestClient().WithAuthFromEnv()
-		res, err := client.Future().InversePerpetual.Balance(bybit.CoinUSDT)
-		{
-			require.NoError(t, err)
-		}
-		{
-			goldenFilename := "./testdata/future-inverse-perpetual/v2-private-wallet-balance.json"
-			testhelper.Compare(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
-			testhelper.UpdateFile(t, goldenFilename, testhelper.ConvertToJSON(res.Result))
-		}
-	})
-
-	t.Run("auth error", func(t *testing.T) {
-		client := bybit.NewTestClient()
-		_, err := client.Future().InversePerpetual.Balance(bybit.CoinBTC)
-		require.Error(t, err)
-	})
-}
