+++ integrationtest/spot-v1/Makefile
@@ -1,0 +1,7 @@
+BYBIT_TEST_UPDATED=false
+
+test:
+	BYBIT_TEST_UPDATED=${BYBIT_TEST_UPDATED} BYBIT_TEST_KEY=${BYBIT_TEST_KEY} BYBIT_TEST_SECRET=${BYBIT_TEST_SECRET} go test github.com/hirokisan/bybit/integrationtest/spot-v1 -v -tags=integrationtestspotv1
+
+test-spec:
+	BYBIT_TEST_UPDATED=${BYBIT_TEST_UPDATED} BYBIT_TEST_KEY=${BYBIT_TEST_KEY} BYBIT_TEST_SECRET=${BYBIT_TEST_SECRET} go test github.com/hirokisan/bybit/integrationtest/spot-v1 -v -tags=integrationtestspotv1 -run ${BYBIT_TEST_METHOD}
