+++ integrationtest/spot-v1/testdata/spot-quote-v1-depth.json
