+++ integrationtest/spot-v1/testdata/spot-quote-v1-kline.json
