+++ integrationtest/spot-v1/testdata/spot-quote-v1-ticker-24hr.json
