+++ integrationtest/spot-v1/testdata/spot-quote-v1-ticker-price.json
