+++ integrationtest/spot-v1/testdata/spot-v1-delete-order.json
