+++ integrationtest/spot-v1/testdata/spot-v1-order-batch-cancel.json
