+++ integrationtest/spot-v1/testdata/spot-v1-symbols.json
