--- integrationtest/testdata/v2-public-mark-price-kline.json
@@ -1,120 +1,0 @@
-{
-  "ret_code": 0,
-  "ret_msg": "OK",
-  "ext_code": "",
-  "ext_info": "",
-  "time_now": "1637762546.081009",
-  "rate_limit_status": 0,
-  "rate_limit_reset_ms": 0,
-  "rate_limit": 0,
-  "result": [
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637683200,
-      "open": 56893.5390625,
-      "high": 57680.37890625,
-      "low": 56893.5390625,
-      "close": 57266.078125
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637690400,
-      "open": 57266.078125,
-      "high": 57840.98828125,
-      "low": 57142.671875,
-      "close": 57798.12890625
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637697600,
-      "open": 57798.12890625,
-      "high": 57880.078125,
-      "low": 57400.55859375,
-      "close": 57704.33984375
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637704800,
-      "open": 57704.33984375,
-      "high": 57753.73046875,
-      "low": 57471.73046875,
-      "close": 57570.80859375
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637712000,
-      "open": 57570.80859375,
-      "high": 57746.69140625,
-      "low": 57046.91015625,
-      "close": 57060.171875
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637719200,
-      "open": 57060.171875,
-      "high": 57136.48046875,
-      "low": 56175.6484375,
-      "close": 56426.37109375
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637726400,
-      "open": 56426.37109375,
-      "high": 56805.03125,
-      "low": 56373.390625,
-      "close": 56621.55859375
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637733600,
-      "open": 56621.55859375,
-      "high": 56740.23046875,
-      "low": 56254.1015625,
-      "close": 56533.23046875
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637740800,
-      "open": 56533.23046875,
-      "high": 57406.2890625,
-      "low": 56431.859375,
-      "close": 56704.140625
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637748000,
-      "open": 56704.140625,
-      "high": 56887.62109375,
-      "low": 56349.55859375,
-      "close": 56648.4609375
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637755200,
-      "open": 56648.4609375,
-      "high": 56876.3515625,
-      "low": 56054.0390625,
-      "close": 56060.23046875
-    },
-    {
-      "symbol": "BTCUSD",
-      "period": "120",
-      "start_at": 1637762400,
-      "open": 56060.23046875,
-      "high": 56309.98046875,
-      "low": 56028.0703125,
-      "close": 56231.30859375
-    }
-  ]
-}
