+++ v2_future_inverse_future.go
@@ -3,4 +3,6 @@ package bybit
 // FutureInverseFutureService :
 type FutureInverseFutureService struct {
 	client *Client
+
+	*FutureCommonService
 }
