+++ v2_future_usdt_perpetual.go
@@ -3,4 +3,6 @@ package bybit
 // FutureUSDTPerpetualService :
 type FutureUSDTPerpetualService struct {
 	client *Client
+
+	*FutureCommonService
 }
